requirements.txt
streamlit
pandas
python-dotenv
sqlalchemy
bcrypt
email-validator
pytz